feat: Prev 활성화, TopBar Component 분리, Detail page Hidden 설정
