+++ src/components/pages/quest/top-bar/index.tsx
@@ -1,0 +1,3 @@
+import { TopBar } from './TopBar';
+
+export { TopBar };
