+++ src/components/template/quest/completed-search/CompletedSearchTemplate.tsx
@@ -1,8 +1,5 @@
 import { Dispatch, SetStateAction } from 'react';
-import { RoundInput } from '@/components/common/round-input';
 import { Swiper as SwiperClass } from 'swiper/types';
-// eslint-disable-next-line import/no-unresolved
-import PrevIcon from '@/assets/images/PrevArrow.svg';
 import { MusicAlbumFrom } from '@/components/pages/quest/music-album-form';
 import { ArtistAvataFrom } from '@/components/pages/quest/artist-avata-form';
 import {
@@ -11,12 +8,14 @@ import {
   getTrackProps,
 } from '@/apis/search';
 import { ArtistDetailForm } from '@/components/pages/quest/artist-detail-from';
+import { TopBar } from '@/components/pages/quest/top-bar';
+
 import * as style from './CompletedSearchTemplate.style';
 import 'swiper/swiper.min.css';
 import 'swiper/swiper-bundle.min.css';
 
 interface CompletedSearchTemplateProps {
-  slideStep: number;
+  isHidden: boolean;
   onSlideChange: (swiper: any) => void;
   completedTextValue: string;
   spotifyArtistForDetailArtist: getSpotifyArtistProps | undefined;
@@ -26,6 +25,7 @@ interface CompletedSearchTemplateProps {
   onClickCategory: (category: string) => void;
   onClickAlbumLikeButton: (trackId: string) => void;
   onClickArtistAvata: (artistData: getSpotifyArtistProps) => void;
+  onClickPrev: () => void;
   switchActiveCategory: number;
   spotifyArtistData: getSpotifyArtistProps[] | undefined;
   spotifyTrackData: getTrackProps[] | undefined;
@@ -33,7 +33,7 @@ interface CompletedSearchTemplateProps {
 }
 
 export const CompletedSearchTemplate = ({
-  slideStep,
+  isHidden,
   onSlideChange,
   completedTextValue,
   spotifyArtistForDetailArtist,
@@ -43,43 +43,23 @@ export const CompletedSearchTemplate = ({
   onClickCategory,
   onClickAlbumLikeButton,
   onClickArtistAvata,
+  onClickPrev,
   switchActiveCategory,
   spotifyArtistData,
   spotifyTrackData,
   commaUserData,
 }: CompletedSearchTemplateProps) => (
   <style.Wrapper>
-    <style.StickySections>
-      <style.TopBox>
-        <PrevIcon />
-        <RoundInput
-          completedTextValue={completedTextValue}
-          onClickRoundInput={onClickRoundInput}
-          handleEraseIconClick={onClickEraseButton}
-          isCompletedSearch
-        />
-      </style.TopBox>
-      <style.ButtonBox>
-        <style.Button
-          isActivate={switchActiveCategory === 0}
-          onClick={() => onClickCategory('music')}
-        >
-          음악
-        </style.Button>
-        <style.Button
-          isActivate={switchActiveCategory === 1}
-          onClick={() => onClickCategory('artist')}
-        >
-          가수
-        </style.Button>
-        <style.Button
-          isActivate={switchActiveCategory === 2}
-          onClick={() => onClickCategory('commaUser')}
-        >
-          사용자
-        </style.Button>
-      </style.ButtonBox>
-    </style.StickySections>
+    <TopBar
+      onClickPrev={onClickPrev}
+      completedTextValue={completedTextValue}
+      onClickRoundInput={onClickRoundInput}
+      onClickEraseButton={onClickEraseButton}
+      onClickCategory={onClickCategory}
+      switchActiveCategory={switchActiveCategory}
+      isHidden={isHidden}
+    />
+
     {switchActiveCategory === 0 && (
       <MusicAlbumFrom
         musicData={spotifyTrackData?.map((value) => ({
